refactor: finish module decomposition and close out refactor plan
