+++ src/cli/parse/helpers.rs
@@ -1,0 +1,23 @@
+use super::super::error::CliError;
+
+pub(super) fn value_as_string(
+    parser: &mut lexopt::Parser,
+    error_message: &'static str,
+) -> Result<String, CliError> {
+    parser
+        .value()?
+        .into_string()
+        .map_err(|_| CliError::message(error_message))
+}
+
+pub(super) fn parse_column_list(
+    parser: &mut lexopt::Parser,
+    error_message: &'static str,
+) -> Result<Vec<usize>, CliError> {
+    let columns = value_as_string(parser, "Column specification must be valid UTF-8")?;
+    columns
+        .split(',')
+        .map(|part| part.trim().parse::<usize>())
+        .collect::<Result<Vec<_>, _>>()
+        .map_err(|_| CliError::message(error_message))
+}
