+++ src/cli/types/mod.rs
@@ -1,0 +1,5 @@
+mod enums;
+mod opts;
+
+pub use enums::{MatchMode, PinnedOrderMode, RankingMode};
+pub use opts::Opts;
